requirements.txt
anthropic
click
python-dotenv
